UIIN-3038 added tests for InventorySettings
